% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_file_size.R
\name{get_one_file_size}
\alias{get_one_file_size}
\title{Obtenir la taille d'un fichier}
\usage{
get_one_file_size(path)
}
\arguments{
\item{path}{A string : path of the file to be analysed.}
}
\value{
df_filesize A \code{data.frame} : name and size of the file pointed by \code{path}.
}
\description{
Obtenir la taille d'un fichier
}
\examples{


# Example with your dataset in "inst/"
datafile <- system.file("nyc_squirrels_sample.csv", package = "fusen")

get_one_file_size(datafile)
# data(sampledf)
# 
# temp_path <- tempfile(pattern = "test_filesize")
# temp_filepath <- file.path(temp_path,"sampledf.csv" )
# write.csv(x = sampledf, temp_filepath) 
# get_one_file_size(sampledf)


}
